[CDEC-457] Combine Pos.Core.Update.Proposal with Pos.Core.Update.Vote
